+++ src/main/java/org/morphix/convert/MapConversions.java
@@ -17,11 +17,11 @@ import java.util.List;
 import java.util.Map;
 
 import org.morphix.convert.function.ConvertFunction;
-import org.morphix.convert.function.PutFunction;
 import org.morphix.convert.function.SimpleConverter;
 import org.morphix.convert.pipeline.MapConversionPipeline;
 import org.morphix.convert.strategy.ConversionStrategy;
 import org.morphix.lang.function.InstanceFunction;
+import org.morphix.lang.function.PutFunction;
 import org.morphix.reflection.ExtendedField;
 
 /**
@@ -92,17 +92,40 @@ public interface MapConversions {
 	 * @param sourceMap source map
 	 * @param keyConverter key converter function
 	 * @param valueConverter value converter function
+	 * @param putFunction put function
+	 * @return destination map
+	 */
+	static <K, S, H, D> MapConversionPipeline<K, S, H, D> convertMap(final Map<K, S> sourceMap, final SimpleConverter<K, H> keyConverter,
+			final SimpleConverter<S, D> valueConverter, final PutFunction<H, D> putFunction) {
+		return new MapConversionPipeline<>(sourceMap, keyConverter, valueConverter, putFunction);
+	}
+
+	/**
+	 * Convenience static method for converting maps.
+	 *
+	 * @param <K> source key type
+	 * @param <S> source value type
+	 * @param <H> destination key type
+	 * @param <D> destination value type
+	 *
+	 * @param sourceMap source map
+	 * @param keyConverter key converter function
+	 * @param valueConverter value converter function
 	 * @return destination map
 	 */
 	static <K, S, H, D> MapConversionPipeline<K, S, H, D> convertMap(final Map<K, S> sourceMap, final SimpleConverter<K, H> keyConverter,
 			final SimpleConverter<S, D> valueConverter) {
-		return new MapConversionPipeline<>(sourceMap, keyConverter, valueConverter);
+		return convertMap(sourceMap, keyConverter, valueConverter, Map::put);
 	}
 
 	/**
 	 * Convenience static method to convert an object to a map conversion pipeline. If the source is null, an empty map is
-	 * returned. The putValueFunction is used to put values into the map and can be used to customize the behavior (e.g., to
+	 * returned. The putFunction is used to put values into the map and can be used to customize the behavior (e.g., to
 	 * handle specific types or to apply transformations or even filtering).
+	 * <p>
+	 * If the source is null, an empty map conversion pipeline is returned.
+	 * <p>
+	 * For sources that are already maps, use the {@link #convertMap(Map, SimpleConverter, SimpleConverter, PutFunction)}.
 	 *
 	 * @param <S> source type
 	 * @param <H> map key type
@@ -111,27 +134,31 @@ public interface MapConversions {
 	 * @param source source object
 	 * @param keyConverter key converter function
 	 * @param valueConverter value converter function
-	 * @param putValueFunction map put function
+	 * @param putFunction map put function
 	 * @return destination map conversion pipeline
 	 * @throws NullPointerException if any of the converters or the map instance function is null
 	 */
 	static <S, H, D> MapConversionPipeline<String, Object, H, D> convert(final S source, final SimpleConverter<String, H> keyConverter,
-			final SimpleConverter<Object, D> valueConverter, final PutFunction<Map<String, Object>, String, Object> putValueFunction) {
+			final SimpleConverter<Object, D> valueConverter, final PutFunction<String, Object> putFunction) {
 		if (source == null) {
 			return convertMap(Map.of(), keyConverter, valueConverter);
 		}
 		List<ExtendedField> fields = ConversionStrategy.findFields(source);
 		Map<String, Object> map = new HashMap<>(fields.size());
 		for (ExtendedField field : fields) {
-			putValueFunction.put(map, field.getName(), field.getFieldValue());
+			putFunction.put(map, field.getName(), field.getFieldValue());
 		}
 		return convertMap(map, keyConverter, valueConverter);
 	}
 
 	/**
 	 * Convenience static method to convert an object to a map. If the source is null, an empty map is returned. The
-	 * putValueFunction is used to put values into the map and can be used to customize the behavior (e.g., to handle
-	 * specific types or to apply transformations or even filtering).
+	 * putFunction is used to put values into the map and can be used to customize the behavior (e.g., to handle specific
+	 * types or to apply transformations or even filtering).
+	 * <p>
+	 * If the source is null, an empty map returned.
+	 * <p>
+	 * For sources that are already maps, use the {@link #convertMap(Map, SimpleConverter, SimpleConverter, PutFunction)}.
 	 *
 	 * @param <S> source type
 	 * @param <H> map key type
@@ -140,16 +167,20 @@ public interface MapConversions {
 	 * @param source source object
 	 * @param keyConverter key converter function
 	 * @param valueConverter value converter function
-	 * @param putValueFunction map put function
+	 * @param putFunction map put function
 	 * @return destination map
 	 */
 	static <S, H, D> Map<H, D> convertToMap(final S source, final SimpleConverter<String, H> keyConverter,
-			final SimpleConverter<Object, D> valueConverter, final PutFunction<Map<String, Object>, String, Object> putValueFunction) {
-		return convert(source, keyConverter, valueConverter, putValueFunction).toMap();
+			final SimpleConverter<Object, D> valueConverter, final PutFunction<String, Object> putFunction) {
+		return convert(source, keyConverter, valueConverter, putFunction).toMap();
 	}
 
 	/**
 	 * Convenience static method to convert an object to a map. If the source is null, an empty map is returned.
+	 * <p>
+	 * If the source is null, an empty map returned.
+	 * <p>
+	 * For sources that are already maps, use the {@link #convertMap(Map, SimpleConverter, SimpleConverter)}.
 	 *
 	 * @param <S> source type
 	 * @param <H> map key type